refactor(infrastructure): consolidate shared persistence under src/infrastructure
Move database, Base model, and session manager from src/config/persistence
to src/infrastructure/persistence, alongside the existing scheduling/ package.
Aligns the layout with CLAUDE.md and ADR-008, which describe src/infrastructure
as the home for cross-module technical infrastructure while src/config stays
focused on settings and DI composition root.
