+++ src/infrastructure/persistence/session_manager.py
