+++ src/modules/library/infrastructure/persistence/models/library_model.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from sqlalchemy import DateTime, String, Text
 from sqlalchemy.orm import Mapped, mapped_column
 
-from src.config.persistence.base import Base
+from src.infrastructure.persistence.base import Base
 
 
 class LibraryModel(Base):
