+++ src/modules/media/infrastructure/persistence/models/media_file.py
@@ -15,7 +15,7 @@ from sqlalchemy import (
 )
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from src.config.persistence.base import Base
+from src.infrastructure.persistence.base import Base
 
 if TYPE_CHECKING:
     from src.modules.media.infrastructure.persistence.models.episode import EpisodeModel
