+++ src/modules/media/infrastructure/persistence/models/series.py
@@ -5,7 +5,7 @@ from typing import TYPE_CHECKING
 from sqlalchemy import Integer, String, Text
 from sqlalchemy.orm import Mapped, mapped_column, relationship
 
-from src.config.persistence.base import Base
+from src.infrastructure.persistence.base import Base
 
 if TYPE_CHECKING:
     from src.modules.media.infrastructure.persistence.models.season import SeasonModel
