+++ src/modules/watch_progress/infrastructure/persistence/models/watch_progress_model.py
@@ -5,7 +5,7 @@ from datetime import datetime
 from sqlalchemy import DateTime, Integer, String
 from sqlalchemy.orm import Mapped, mapped_column
 
-from src.config.persistence.base import Base
+from src.infrastructure.persistence.base import Base
 
 
 class WatchProgressModel(Base):
