+++ tests/modules/library/integration/conftest.py
@@ -5,7 +5,7 @@ from collections.abc import AsyncGenerator
 import pytest
 from sqlalchemy.ext.asyncio import AsyncSession, async_sessionmaker, create_async_engine
 
-from src.config.persistence import Base
+from src.infrastructure.persistence import Base
 
 
 @pytest.fixture(scope="function")
